correction on Section6.3 (Data Studio)
As pointed out by Mike Harris <mike@dynamo6.com>
```
you can turn the prefetch cache oﬀ for a given report. You might want to do this if:
 your data changes frequently and you want to prioritize freshness over performance.
you are using a data source that incurs usage costs (e.g., BigQuery) and want to minimize those costs.
Surely that last point is wrong. You would want the prefetch cache turned on if that was the case, not off.
```